SkyDocu 1.4 dev

+++ app/core/FileManager.php
@@ -122,7 +122,9 @@ class FileManager {
         if(is_dir($dirPath)) {
             return true;
         }
-        return mkdir($dirPath, 0777, $recursive);
+        if(!FileManager::folderExists($dirPath)) {
+            return mkdir($dirPath, 0777, $recursive);
+        }
     }
 
     /**
@@ -154,7 +156,11 @@ class FileManager {
             }
 
             if(!$root) {
-                return rmdir($dirPath) && $result;
+                if(self::folderExists($dirPath)) {
+                    return rmdir($dirPath) && $result;
+                } else {
+                    return $result;
+                }
             } else {
                 return $result;
             }
@@ -185,6 +191,25 @@ class FileManager {
      */
     public static function moveFile(string $oldPath, string $newPath) {
         return rename($oldPath, $newPath);
+    }
+
+    /**
+     * Returns filename from given path
+     * 
+     * @param string $path Path to the file
+     * @param bool $returnExtension True if file extension should be returned as well
+     * @return string Filename
+     */
+    public static function getFilenameFromPath(string $path, bool $returnExtension = false): string {
+        $parts = explode('\\', $path);
+
+        $filenameWithExtension = $parts[count($parts) - 1];
+
+        if($returnExtension) {
+            return $filenameWithExtension;
+        } else {
+            return explode('.', $filenameWithExtension)[0];
+        }
     }
 }
